Break up ext_proc_integration_test (#41256)
This is large but trivial PR that breaks up the
ext_proc_integration_test into smaller suites. Adding new tests to the
same file is unsustainable.

Risk Level: none
Testing: unit tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a
