+++ test/extensions/filters/http/ext_proc/ext_proc_integration_test.cc
@@ -18,6 +18,7 @@
 
 #include "test/common/grpc/grpc_client_integration.h"
 #include "test/common/http/common.h"
+#include "test/extensions/filters/http/ext_proc/ext_proc_integration_common.h"
 #include "test/extensions/filters/http/ext_proc/logging_test_filter.pb.h"
 #include "test/extensions/filters/http/ext_proc/logging_test_filter.pb.validate.h"
 #include "test/extensions/filters/http/ext_proc/tracer_test_filter.pb.h"
@@ -34,6 +35,9 @@
 #include "ocpdiag/core/testing/status_matchers.h"
 
 namespace Envoy {
+namespace Extensions {
+namespace HttpFilters {
+namespace ExternalProcessing {
 
 using envoy::config::route::v3::Route;
 using envoy::config::route::v3::VirtualHost;
@@ -65,880 +69,6 @@ using testing::_;
 using testing::Not;
 
 using namespace std::chrono_literals;
-
-struct ConfigOptions {
-  bool valid_grpc_server = true;
-  bool add_logging_filter = false;
-  absl::optional<LoggingTestFilterConfig> logging_filter_config = absl::nullopt;
-  bool http1_codec = false;
-  bool add_metadata = false;
-  bool downstream_filter = true;
-  bool add_response_processor = false;
-};
-
-// A filter that sticks dynamic metadata info into headers for integration testing.
-class DynamicMetadataToHeadersFilter : public Http::PassThroughFilter {
-public:
-  constexpr static char name[] = "dynamic-metadata-to-headers-filter";
-
-  Http::FilterHeadersStatus decodeHeaders(Http::RequestHeaderMap&, bool) override {
-    return Http::FilterHeadersStatus::Continue;
-  }
-
-  Http::FilterHeadersStatus encodeHeaders(Http::ResponseHeaderMap& headers, bool) override {
-    if (decoder_callbacks_->streamInfo().dynamicMetadata().filter_metadata_size() > 0) {
-      const auto& md = decoder_callbacks_->streamInfo().dynamicMetadata().filter_metadata();
-      for (const auto& md_entry : md) {
-        std::string key_prefix = md_entry.first;
-        for (const auto& field : md_entry.second.fields()) {
-          headers.addCopy(Http::LowerCaseString(key_prefix), field.first);
-        }
-      }
-    }
-    return Http::FilterHeadersStatus::Continue;
-  }
-};
-
-// These tests exercise the ext_proc filter through Envoy's integration test
-// environment by configuring an instance of the Envoy server and driving it
-// through the mock network stack.
-class ExtProcIntegrationTest : public HttpIntegrationTest,
-                               public Grpc::GrpcClientIntegrationParamTest {
-protected:
-  ExtProcIntegrationTest() : HttpIntegrationTest(Http::CodecType::HTTP2, ipVersion()) {}
-
-  void createUpstreams() override {
-    HttpIntegrationTest::createUpstreams();
-
-    // Create separate "upstreams" for ExtProc gRPC servers
-    for (int i = 0; i < grpc_upstream_count_; ++i) {
-      grpc_upstreams_.push_back(&addFakeUpstream(Http::CodecType::HTTP2));
-    }
-  }
-
-  void TearDown() override {
-    if (processor_connection_) {
-      ASSERT_TRUE(processor_connection_->close());
-      ASSERT_TRUE(processor_connection_->waitForDisconnect());
-    }
-    cleanupUpstreamAndDownstream();
-  }
-
-  void initializeConfig(ConfigOptions config_option = {},
-                        const std::vector<std::pair<int, int>>& cluster_endpoints = {{0, 1},
-                                                                                     {1, 1}}) {
-    int total_cluster_endpoints = 0;
-    std::for_each(
-        cluster_endpoints.begin(), cluster_endpoints.end(),
-        [&total_cluster_endpoints](const auto& item) { total_cluster_endpoints += item.second; });
-    ASSERT_EQ(total_cluster_endpoints, grpc_upstream_count_);
-
-    config_helper_.addConfigModifier([this, cluster_endpoints, config_option](
-                                         envoy::config::bootstrap::v3::Bootstrap& bootstrap) {
-      // Ensure "HTTP2 with no prior knowledge." Necessary for gRPC and for headers
-      ConfigHelper::setHttp2(
-          *(bootstrap.mutable_static_resources()->mutable_clusters()->Mutable(0)));
-
-      // Clusters for ExtProc gRPC servers, starting by copying an existing
-      // cluster
-      for (const auto& [cluster_id, endpoint_count] : cluster_endpoints) {
-        auto* server_cluster = bootstrap.mutable_static_resources()->add_clusters();
-        server_cluster->MergeFrom(bootstrap.static_resources().clusters()[0]);
-        std::string cluster_name = absl::StrCat("ext_proc_server_", cluster_id);
-        server_cluster->set_name(cluster_name);
-        server_cluster->mutable_load_assignment()->set_cluster_name(cluster_name);
-        ASSERT_EQ(server_cluster->load_assignment().endpoints_size(), 1);
-        auto* endpoints = server_cluster->mutable_load_assignment()->mutable_endpoints(0);
-        ASSERT_EQ(endpoints->lb_endpoints_size(), 1);
-        for (int i = 1; i < endpoint_count; ++i) {
-          auto* new_lb_endpoint = endpoints->add_lb_endpoints();
-          new_lb_endpoint->MergeFrom(endpoints->lb_endpoints(0));
-        }
-      }
-
-      const std::string valid_grpc_cluster_name = "ext_proc_server_0";
-      if (config_option.valid_grpc_server) {
-        // Load configuration of the server from YAML and use a helper to add a grpc_service
-        // stanza pointing to the cluster that we just made
-        setGrpcService(*proto_config_.mutable_grpc_service(), valid_grpc_cluster_name,
-                       grpc_upstreams_[0]->localAddress());
-      } else {
-        // Set up the gRPC service with wrong cluster name and address.
-        setGrpcService(*proto_config_.mutable_grpc_service(), "ext_proc_wrong_server",
-                       std::make_shared<Network::Address::Ipv4Instance>("127.0.0.1", 1234));
-      }
-
-      std::string ext_proc_filter_name = "envoy.filters.http.ext_proc";
-      if (config_option.downstream_filter) {
-        // Construct a configuration proto for our filter and then re-write it
-        // to JSON so that we can add it to the overall config
-        envoy::extensions::filters::network::http_connection_manager::v3::HttpFilter
-            ext_proc_filter;
-        ext_proc_filter.set_name(ext_proc_filter_name);
-        ext_proc_filter.mutable_typed_config()->PackFrom(proto_config_);
-        config_helper_.prependFilter(MessageUtil::getJsonStringFromMessageOrError(ext_proc_filter));
-      }
-
-      // Add set_metadata filter to inject dynamic metadata used for testing
-      if (config_option.add_metadata) {
-        envoy::config::listener::v3::Filter set_metadata_filter;
-        std::string set_metadata_filter_name = "envoy.filters.http.set_metadata";
-        set_metadata_filter.set_name(set_metadata_filter_name);
-
-        envoy::extensions::filters::http::set_metadata::v3::Config set_metadata_config;
-        auto* untyped_md = set_metadata_config.add_metadata();
-        untyped_md->set_metadata_namespace("forwarding_ns_untyped");
-        untyped_md->set_allow_overwrite(true);
-        Protobuf::Struct test_md_val;
-        (*test_md_val.mutable_fields())["foo"].set_string_value("value from set_metadata");
-        (*untyped_md->mutable_value()) = test_md_val;
-
-        auto* typed_md = set_metadata_config.add_metadata();
-        typed_md->set_metadata_namespace("forwarding_ns_typed");
-        typed_md->set_allow_overwrite(true);
-        envoy::extensions::filters::http::set_metadata::v3::Metadata typed_md_to_stuff;
-        typed_md_to_stuff.set_metadata_namespace("typed_value from set_metadata");
-        typed_md->mutable_typed_value()->PackFrom(typed_md_to_stuff);
-
-        set_metadata_filter.mutable_typed_config()->PackFrom(set_metadata_config);
-        config_helper_.prependFilter(
-            MessageUtil::getJsonStringFromMessageOrError(set_metadata_filter));
-
-        // Add filter that dumps streamInfo into headers so we can check our receiving
-        // namespaces
-        config_helper_.prependFilter(fmt::format(R"EOF(
-  name: stream-info-to-headers-filter
-        )EOF"));
-      }
-
-      // Add dynamic_metadata_to_headers filter to inject dynamic metadata used for testing
-      if (config_option.add_response_processor) {
-        simple_filter_config_ =
-            std::make_unique<SimpleFilterConfig<DynamicMetadataToHeadersFilter>>();
-        registration_ = std::make_unique<
-            Envoy::Registry::InjectFactory<Server::Configuration::NamedHttpFilterConfigFactory>>(
-            *simple_filter_config_);
-        config_helper_.prependFilter(fmt::format(R"EOF(
-         name: dynamic-metadata-to-headers-filter
-         typed_config:
-           "@type": type.googleapis.com/google.protobuf.Struct
-        )EOF"));
-      }
-
-      // Add logging test filter only in Envoy gRPC mode.
-      // gRPC side stream logging is only supported in Envoy gRPC mode at the moment.
-      if (clientType() == Grpc::ClientType::EnvoyGrpc && config_option.add_logging_filter &&
-          config_option.valid_grpc_server) {
-        LoggingTestFilterConfig logging_filter_config;
-        logging_filter_config.set_logging_id(ext_proc_filter_name);
-        logging_filter_config.set_upstream_cluster_name(valid_grpc_cluster_name);
-        // No need to check the bytes received for observability mode because it is a
-        // "send and go" mode.
-        logging_filter_config.set_check_received_bytes(!proto_config_.observability_mode());
-        logging_filter_config.set_http_rcd("via_upstream");
-        if (config_option.logging_filter_config) {
-          logging_filter_config.MergeFrom(*config_option.logging_filter_config);
-        }
-
-        envoy::extensions::filters::network::http_connection_manager::v3::HttpFilter logging_filter;
-        logging_filter.set_name("logging-test-filter");
-        logging_filter.mutable_typed_config()->PackFrom(logging_filter_config);
-
-        config_helper_.prependFilter(MessageUtil::getJsonStringFromMessageOrError(logging_filter));
-      }
-
-      // Parameterize with defer processing to prevent bit rot as filter made
-      // assumptions of data flow, prior relying on eager processing.
-    });
-
-    if (config_option.add_response_processor) {
-      processing_response_factory_ = std::make_unique<TestOnProcessingResponseFactory>();
-      processing_response_factory_registration_ =
-          std::make_unique<Envoy::Registry::InjectFactory<OnProcessingResponseFactory>>(
-              *processing_response_factory_);
-      Protobuf::Struct config;
-      proto_config_.mutable_on_processing_response()->set_name("test-on-processing-response");
-      proto_config_.mutable_on_processing_response()->mutable_typed_config()->PackFrom(config);
-    }
-
-    if (config_option.http1_codec) {
-      setUpstreamProtocol(Http::CodecType::HTTP1);
-      setDownstreamProtocol(Http::CodecType::HTTP1);
-    } else {
-      setUpstreamProtocol(Http::CodecType::HTTP2);
-      setDownstreamProtocol(Http::CodecType::HTTP2);
-    }
-  }
-
-  bool IsEnvoyGrpc() { return std::get<1>(GetParam()) == Envoy::Grpc::ClientType::EnvoyGrpc; }
-
-  void setPerRouteConfig(Route* route, const ExtProcPerRoute& cfg) {
-    Any cfg_any;
-    ASSERT_TRUE(cfg_any.PackFrom(cfg));
-    route->mutable_typed_per_filter_config()->insert(
-        MapPair<std::string, Any>("envoy.filters.http.ext_proc", cfg_any));
-  }
-
-  void setPerHostConfig(VirtualHost& vh, const ExtProcPerRoute& cfg) {
-    Any cfg_any;
-    ASSERT_TRUE(cfg_any.PackFrom(cfg));
-    vh.mutable_typed_per_filter_config()->insert(
-        MapPair<std::string, Any>("envoy.filters.http.ext_proc", cfg_any));
-  }
-
-  void protocolConfigEncoding(ProcessingRequest& request) {
-    protocol_config_encoded_ = false;
-    if (request.has_protocol_config()) {
-      protocol_config_encoded_ = true;
-      protocol_config_ = request.protocol_config();
-    }
-  }
-
-  IntegrationStreamDecoderPtr sendDownstreamRequest(
-      absl::optional<std::function<void(Http::RequestHeaderMap& headers)>> modify_headers) {
-    auto conn = makeClientConnection(lookupPort("http"));
-    codec_client_ = makeHttpConnection(std::move(conn));
-    Http::TestRequestHeaderMapImpl headers;
-    HttpTestUtility::addDefaultHeaders(headers);
-    if (modify_headers) {
-      (*modify_headers)(headers);
-    }
-    return codec_client_->makeHeaderOnlyRequest(headers);
-  }
-
-  IntegrationStreamDecoderPtr sendDownstreamRequestWithBody(
-      absl::string_view body,
-      absl::optional<std::function<void(Http::RequestHeaderMap& headers)>> modify_headers,
-      bool add_content_length = false) {
-    auto conn = makeClientConnection(lookupPort("http"));
-    codec_client_ = makeHttpConnection(std::move(conn));
-    Http::TestRequestHeaderMapImpl headers;
-    HttpTestUtility::addDefaultHeaders(headers);
-    headers.setMethod("POST");
-    if (modify_headers) {
-      (*modify_headers)(headers);
-    }
-
-    if (add_content_length) {
-      headers.setContentLength(body.size());
-    }
-    return codec_client_->makeRequestWithBody(headers, std::string(body));
-  }
-
-  IntegrationStreamDecoderPtr sendDownstreamRequestWithBodyAndTrailer(absl::string_view body) {
-    codec_client_ = makeHttpConnection(lookupPort("http"));
-    Http::TestRequestHeaderMapImpl headers;
-    HttpTestUtility::addDefaultHeaders(headers);
-
-    auto encoder_decoder = codec_client_->startRequest(headers);
-    request_encoder_ = &encoder_decoder.first;
-    auto response = std::move(encoder_decoder.second);
-    codec_client_->sendData(*request_encoder_, body, false);
-    Http::TestRequestTrailerMapImpl request_trailers{{"x-trailer-foo", "yes"}};
-    codec_client_->sendTrailers(*request_encoder_, request_trailers);
-
-    return response;
-  }
-
-  void verifyDownstreamResponse(IntegrationStreamDecoder& response, int status_code) {
-    ASSERT_TRUE(response.waitForEndStream());
-    EXPECT_TRUE(response.complete());
-    EXPECT_EQ(std::to_string(status_code), response.headers().getStatusValue());
-  }
-
-  void handleUpstreamRequest(bool add_content_length = false, int status_code = 200) {
-    ASSERT_TRUE(fake_upstreams_[0]->waitForHttpConnection(*dispatcher_, fake_upstream_connection_));
-    ASSERT_TRUE(fake_upstream_connection_->waitForNewStream(*dispatcher_, upstream_request_));
-    ASSERT_TRUE(upstream_request_->waitForEndStream(*dispatcher_));
-    Http::TestResponseHeaderMapImpl response_headers =
-        Http::TestResponseHeaderMapImpl{{":status", std::to_string(status_code)}};
-    uint64_t content_length = 100;
-    if (add_content_length) {
-      response_headers.setContentLength(content_length);
-    }
-    upstream_request_->encodeHeaders(response_headers, false);
-    upstream_request_->encodeData(content_length, true);
-  }
-
-  void verifyChunkedEncoding(const Http::RequestOrResponseHeaderMap& headers) {
-    EXPECT_EQ(headers.ContentLength(), nullptr);
-    EXPECT_THAT(headers, ContainsHeader(Http::Headers::get().TransferEncoding,
-                                        Http::Headers::get().TransferEncodingValues.Chunked));
-  }
-
-  void handleUpstreamRequestWithTrailer() {
-    ASSERT_TRUE(fake_upstreams_[0]->waitForHttpConnection(*dispatcher_, fake_upstream_connection_));
-    ASSERT_TRUE(fake_upstream_connection_->waitForNewStream(*dispatcher_, upstream_request_));
-    ASSERT_TRUE(upstream_request_->waitForEndStream(*dispatcher_));
-    upstream_request_->encodeHeaders(Http::TestResponseHeaderMapImpl{{":status", "200"}}, false);
-    upstream_request_->encodeData(100, false);
-    upstream_request_->encodeTrailers(Http::TestResponseTrailerMapImpl{{"x-test-trailers", "Yes"}});
-  }
-
-  void handleUpstreamRequestWithResponse(const Buffer::Instance& all_data, uint64_t chunk_size) {
-    ASSERT_TRUE(fake_upstreams_[0]->waitForHttpConnection(*dispatcher_, fake_upstream_connection_));
-    ASSERT_TRUE(fake_upstream_connection_->waitForNewStream(*dispatcher_, upstream_request_));
-    ASSERT_TRUE(upstream_request_->waitForEndStream(*dispatcher_));
-    upstream_request_->encodeHeaders(Http::TestResponseHeaderMapImpl{{":status", "200"}}, false);
-
-    // Copy the data so that we don't modify it
-    Buffer::OwnedImpl total_response = all_data;
-    while (total_response.length() > 0) {
-      auto to_move = std::min(total_response.length(), chunk_size);
-      Buffer::OwnedImpl chunk;
-      chunk.move(total_response, to_move);
-      EXPECT_EQ(to_move, chunk.length());
-      upstream_request_->encodeData(chunk, false);
-    }
-    upstream_request_->encodeData(0, true);
-  }
-
-  void waitForFirstMessage(FakeUpstream& grpc_upstream, ProcessingRequest& request) {
-    ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-    ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-  }
-
-  void processGenericMessage(
-      FakeUpstream& grpc_upstream, bool first_message,
-      absl::optional<std::function<bool(const ProcessingRequest&, ProcessingResponse&)>> cb) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response;
-    const bool sendReply = !cb || (*cb)(request, response);
-    if (sendReply) {
-      processor_stream_->sendGrpcMessage(response);
-    }
-  }
-
-  void processRequestHeadersMessage(
-      FakeUpstream& grpc_upstream, bool first_message,
-      absl::optional<std::function<bool(const HttpHeaders&, HeadersResponse&)>> cb) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    ASSERT_TRUE(request.has_request_headers());
-    protocolConfigEncoding(request);
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response;
-    auto* headers = response.mutable_request_headers();
-    const bool sendReply = !cb || (*cb)(request.request_headers(), *headers);
-    if (sendReply) {
-      processor_stream_->sendGrpcMessage(response);
-    }
-  }
-
-  void processRequestTrailersMessage(
-      FakeUpstream& grpc_upstream, bool first_message,
-      absl::optional<std::function<bool(const HttpTrailers&, TrailersResponse&)>> cb) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    ASSERT_TRUE(request.has_request_trailers());
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response;
-    auto* body = response.mutable_request_trailers();
-    const bool sendReply = !cb || (*cb)(request.request_trailers(), *body);
-    if (sendReply) {
-      processor_stream_->sendGrpcMessage(response);
-    }
-  }
-
-  void processResponseHeadersMessage(
-      FakeUpstream& grpc_upstream, bool first_message,
-      absl::optional<std::function<bool(const HttpHeaders&, HeadersResponse&)>> cb) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    ASSERT_TRUE(request.has_response_headers());
-    protocolConfigEncoding(request);
-
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response;
-    auto* headers = response.mutable_response_headers();
-    const bool sendReply = !cb || (*cb)(request.response_headers(), *headers);
-    if (sendReply) {
-      processor_stream_->sendGrpcMessage(response);
-    }
-  }
-
-  void
-  processRequestBodyMessage(FakeUpstream& grpc_upstream, bool first_message,
-                            absl::optional<std::function<bool(const HttpBody&, BodyResponse&)>> cb,
-                            bool check_downstream_flow_control = false) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    ASSERT_TRUE(request.has_request_body());
-    protocolConfigEncoding(request);
-
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-
-    if (check_downstream_flow_control) {
-      // Check the flow control counter in downstream, which is triggered on the request
-      // path to ext_proc server (i.e., from side stream).
-      test_server_->waitForCounterGe(
-          "http.config_test.downstream_flow_control_paused_reading_total", 1);
-    }
-
-    // Send back the response from ext_proc server.
-    ProcessingResponse response;
-    auto* body = response.mutable_request_body();
-    const bool sendReply = !cb || (*cb)(request.request_body(), *body);
-    if (sendReply) {
-      processor_stream_->sendGrpcMessage(response);
-    }
-  }
-
-  void processResponseBodyMessage(
-      FakeUpstream& grpc_upstream, bool first_message,
-      absl::optional<std::function<bool(const HttpBody&, BodyResponse&)>> cb) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    ASSERT_TRUE(request.has_response_body());
-    protocolConfigEncoding(request);
-
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response;
-    auto* body = response.mutable_response_body();
-    const bool sendReply = !cb || (*cb)(request.response_body(), *body);
-    if (sendReply) {
-      processor_stream_->sendGrpcMessage(response);
-    }
-  }
-
-  void processResponseTrailersMessage(
-      FakeUpstream& grpc_upstream, bool first_message,
-      absl::optional<std::function<bool(const HttpTrailers&, TrailersResponse&)>> cb) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    ASSERT_TRUE(request.has_response_trailers());
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response;
-    auto* body = response.mutable_response_trailers();
-    const bool sendReply = !cb || (*cb)(request.response_trailers(), *body);
-    if (sendReply) {
-      processor_stream_->sendGrpcMessage(response);
-    }
-  }
-
-  void processAndRespondImmediately(FakeUpstream& grpc_upstream, bool first_message,
-                                    absl::optional<std::function<void(ImmediateResponse&)>> cb) {
-    ProcessingRequest request;
-    if (first_message) {
-      ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-      ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response;
-    auto* immediate = response.mutable_immediate_response();
-    if (cb) {
-      (*cb)(*immediate);
-    }
-    processor_stream_->sendGrpcMessage(response);
-  }
-
-  // ext_proc server sends back a response to tell Envoy to stop the
-  // original timer and start a new timer.
-  void serverSendNewTimeout(const uint64_t timeout_ms) {
-    ProcessingResponse response;
-    if (timeout_ms < 1000) {
-      response.mutable_override_message_timeout()->set_nanos(timeout_ms * 1000000);
-    } else {
-      response.mutable_override_message_timeout()->set_seconds(timeout_ms / 1000);
-    }
-    processor_stream_->sendGrpcMessage(response);
-  }
-
-  // The new timeout message is ignored by Envoy due to different reasons, like
-  // new_timeout setting is out-of-range, or max_message_timeout is not configured.
-  void newTimeoutWrongConfigTest(const uint64_t timeout_ms) {
-    // Set envoy filter timeout to be 200ms.
-    proto_config_.mutable_message_timeout()->set_nanos(200000000);
-    // Config max_message_timeout proto to enable the new timeout API.
-    if (max_message_timeout_ms_) {
-      if (max_message_timeout_ms_ < 1000) {
-        proto_config_.mutable_max_message_timeout()->set_nanos(max_message_timeout_ms_ * 1000000);
-      } else {
-        proto_config_.mutable_max_message_timeout()->set_seconds(max_message_timeout_ms_ / 1000);
-      }
-    }
-    initializeConfig();
-    HttpIntegrationTest::initialize();
-    auto response = sendDownstreamRequest(absl::nullopt);
-
-    processRequestHeadersMessage(*grpc_upstreams_[0], true,
-                                 [&](const HttpHeaders&, HeadersResponse&) {
-                                   serverSendNewTimeout(timeout_ms);
-                                   // ext_proc server stays idle for 300ms before sending back the
-                                   // response.
-                                   timeSystem().advanceTimeWaitImpl(300ms);
-                                   return true;
-                                 });
-    // Verify the new timer is not started and the original timer timeouts,
-    // and downstream receives 504.
-    verifyDownstreamResponse(*response, 504);
-  }
-
-  void addMutationSetHeaders(const int count,
-                             envoy::service::ext_proc::v3::HeaderMutation& mutation) {
-    for (int i = 0; i < count; i++) {
-      auto* headers = mutation.add_set_headers();
-      auto str = absl::StrCat("x-test-header-internal-", std::to_string(i));
-      headers->mutable_append()->set_value(false);
-      headers->mutable_header()->set_key(str);
-      headers->mutable_header()->set_raw_value(str);
-    }
-  }
-
-  // Verify content-length header set by external processor is removed and chunked encoding is
-  // enabled.
-  void testWithHeaderMutation(ConfigOptions config_option) {
-    initializeConfig(config_option);
-    HttpIntegrationTest::initialize();
-
-    auto response = sendDownstreamRequestWithBody("Replace this!", absl::nullopt);
-    processRequestHeadersMessage(
-        *grpc_upstreams_[0], true, [](const HttpHeaders&, HeadersResponse& headers_resp) {
-          auto* content_length =
-              headers_resp.mutable_response()->mutable_header_mutation()->add_set_headers();
-          content_length->mutable_append()->set_value(false);
-          content_length->mutable_header()->set_key("content-length");
-          content_length->mutable_header()->set_raw_value("13");
-          return true;
-        });
-
-    processRequestBodyMessage(
-        *grpc_upstreams_[0], false, [](const HttpBody& body, BodyResponse& body_resp) {
-          EXPECT_TRUE(body.end_of_stream());
-          auto* body_mut = body_resp.mutable_response()->mutable_body_mutation();
-          body_mut->set_body("Hello, World!");
-          return true;
-        });
-    handleUpstreamRequest();
-    // Verify that the content length header is removed and chunked encoding is enabled by http1
-    // codec.
-    verifyChunkedEncoding(upstream_request_->headers());
-
-    EXPECT_EQ(upstream_request_->body().toString(), "Hello, World!");
-    verifyDownstreamResponse(*response, 200);
-  }
-
-  // Verify existing content-length header (i.e., no external processor mutation) is removed and
-  // chunked encoding is enabled.
-  void testWithoutHeaderMutation(ConfigOptions config_option) {
-    initializeConfig(config_option);
-    HttpIntegrationTest::initialize();
-
-    auto response =
-        sendDownstreamRequestWithBody("test!", absl::nullopt, /*add_content_length=*/true);
-    processRequestHeadersMessage(*grpc_upstreams_[0], true, absl::nullopt);
-    processRequestBodyMessage(
-        *grpc_upstreams_[0], false, [](const HttpBody& body, BodyResponse& body_resp) {
-          EXPECT_TRUE(body.end_of_stream());
-          auto* body_mut = body_resp.mutable_response()->mutable_body_mutation();
-          body_mut->set_body("Hello, World!");
-          return true;
-        });
-
-    handleUpstreamRequest();
-    verifyChunkedEncoding(upstream_request_->headers());
-
-    EXPECT_EQ(upstream_request_->body().toString(), "Hello, World!");
-    verifyDownstreamResponse(*response, 200);
-  }
-
-  void addMutationRemoveHeaders(const int count,
-                                envoy::service::ext_proc::v3::HeaderMutation& mutation) {
-    for (int i = 0; i < count; i++) {
-      mutation.add_remove_headers(absl::StrCat("x-test-header-internal-", std::to_string(i)));
-    }
-  }
-
-  void testGetAndFailStream() {
-    HttpIntegrationTest::initialize();
-    auto response = sendDownstreamRequest(absl::nullopt);
-
-    ProcessingRequest request_headers_msg;
-    waitForFirstMessage(*grpc_upstreams_[0], request_headers_msg);
-    // Fail the stream immediately
-    processor_stream_->encodeHeaders(Http::TestResponseHeaderMapImpl{{":status", "500"}}, true);
-    verifyDownstreamResponse(*response, 500);
-  }
-
-  void testGetAndCloseStream() {
-    HttpIntegrationTest::initialize();
-    auto response = sendDownstreamRequest(absl::nullopt);
-
-    ProcessingRequest request_headers_msg;
-    waitForFirstMessage(*grpc_upstreams_[0], request_headers_msg);
-    // Just close the stream without doing anything
-    processor_stream_->startGrpcStream();
-    processor_stream_->finishGrpcStream(Grpc::Status::Ok);
-
-    handleUpstreamRequest();
-    verifyDownstreamResponse(*response, 200);
-  }
-
-  void testSendDyanmicMetadata() {
-    Protobuf::Struct test_md_struct;
-    (*test_md_struct.mutable_fields())["foo"].set_string_value("value from ext_proc");
-
-    Protobuf::Value md_val;
-    *(md_val.mutable_struct_value()) = test_md_struct;
-
-    processGenericMessage(
-        *grpc_upstreams_[0], true,
-        [md_val](const ProcessingRequest& req, ProcessingResponse& resp) {
-          // Verify the processing request contains the untyped metadata we injected.
-          EXPECT_TRUE(req.metadata_context().filter_metadata().contains("forwarding_ns_untyped"));
-          const Protobuf::Struct& fwd_metadata =
-              req.metadata_context().filter_metadata().at("forwarding_ns_untyped");
-          EXPECT_EQ(1, fwd_metadata.fields_size());
-          EXPECT_TRUE(fwd_metadata.fields().contains("foo"));
-          EXPECT_EQ("value from set_metadata", fwd_metadata.fields().at("foo").string_value());
-
-          // Verify the processing request contains the typed metadata we injected.
-          EXPECT_TRUE(
-              req.metadata_context().typed_filter_metadata().contains("forwarding_ns_typed"));
-          const Protobuf::Any& fwd_typed_metadata =
-              req.metadata_context().typed_filter_metadata().at("forwarding_ns_typed");
-          EXPECT_EQ("type.googleapis.com/envoy.extensions.filters.http.set_metadata.v3.Metadata",
-                    fwd_typed_metadata.type_url());
-          envoy::extensions::filters::http::set_metadata::v3::Metadata typed_md_from_req;
-          fwd_typed_metadata.UnpackTo(&typed_md_from_req);
-          EXPECT_EQ("typed_value from set_metadata", typed_md_from_req.metadata_namespace());
-
-          // Spoof the response to contain receiving metadata.
-          HeadersResponse headers_resp;
-          (*resp.mutable_request_headers()) = headers_resp;
-          auto mut_md_fields = resp.mutable_dynamic_metadata()->mutable_fields();
-          (*mut_md_fields).emplace("receiving_ns_untyped", md_val);
-
-          return true;
-        });
-  }
-
-  void testSidestreamPushbackDownstream(uint32_t body_size, bool check_downstream_flow_control) {
-    config_helper_.setBufferLimits(1024, 1024);
-
-    proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SKIP);
-    proto_config_.mutable_processing_mode()->set_request_body_mode(ProcessingMode::STREAMED);
-    proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-    initializeConfig();
-    HttpIntegrationTest::initialize();
-
-    std::string body_str = std::string(body_size, 'a');
-    auto response = sendDownstreamRequestWithBody(body_str, absl::nullopt);
-
-    bool end_stream = false;
-    int count = 0;
-    while (!end_stream) {
-      processRequestBodyMessage(
-          *grpc_upstreams_[0], count == 0 ? true : false,
-          [&end_stream](const HttpBody& body, BodyResponse&) {
-            end_stream = body.end_of_stream();
-            return true;
-          },
-          check_downstream_flow_control);
-      count++;
-    }
-    handleUpstreamRequest();
-
-    verifyDownstreamResponse(*response, 200);
-  }
-
-  void initializeConfigDuplexStreamed(bool both_direction = false) {
-    config_helper_.setBufferLimits(1024, 1024);
-    auto* processing_mode = proto_config_.mutable_processing_mode();
-    processing_mode->set_request_header_mode(ProcessingMode::SEND);
-    processing_mode->set_request_body_mode(ProcessingMode::FULL_DUPLEX_STREAMED);
-    processing_mode->set_request_trailer_mode(ProcessingMode::SEND);
-    if (!both_direction) {
-      processing_mode->set_response_header_mode(ProcessingMode::SKIP);
-    } else {
-      processing_mode->set_response_header_mode(ProcessingMode::SEND);
-      processing_mode->set_response_body_mode(ProcessingMode::FULL_DUPLEX_STREAMED);
-      processing_mode->set_response_trailer_mode(ProcessingMode::SEND);
-    }
-
-    initializeConfig();
-    HttpIntegrationTest::initialize();
-  }
-
-  IntegrationStreamDecoderPtr initAndSendDataDuplexStreamedMode(absl::string_view body_sent,
-                                                                bool end_of_stream,
-                                                                bool both_direction = false) {
-    initializeConfigDuplexStreamed(both_direction);
-    codec_client_ = makeHttpConnection(lookupPort("http"));
-    Http::TestRequestHeaderMapImpl default_headers;
-    HttpTestUtility::addDefaultHeaders(default_headers);
-
-    std::pair<Http::RequestEncoder&, IntegrationStreamDecoderPtr> encoder_decoder =
-        codec_client_->startRequest(default_headers);
-    request_encoder_ = &encoder_decoder.first;
-    IntegrationStreamDecoderPtr response = std::move(encoder_decoder.second);
-    codec_client_->sendData(*request_encoder_, body_sent, end_of_stream);
-    return response;
-  }
-
-  void serverReceiveHeaderDuplexStreamed(ProcessingRequest& header, bool first_message = true,
-                                         bool response = false) {
-    if (first_message) {
-      EXPECT_TRUE(grpc_upstreams_[0]->waitForHttpConnection(*dispatcher_, processor_connection_));
-      EXPECT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    }
-    EXPECT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, header));
-    if (response) {
-      EXPECT_TRUE(header.has_response_headers());
-    } else {
-      EXPECT_TRUE(header.has_request_headers());
-    }
-  }
-
-  uint32_t serverReceiveBodyDuplexStreamed(absl::string_view body_sent, bool response = false,
-                                           bool compare_body = true) {
-    std::string body_received;
-    bool end_stream = false;
-    uint32_t total_req_body_msg = 0;
-    while (!end_stream) {
-      ProcessingRequest body_request;
-      EXPECT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, body_request));
-      if (response) {
-        EXPECT_TRUE(body_request.has_response_body());
-        body_received = absl::StrCat(body_received, body_request.response_body().body());
-        end_stream = body_request.response_body().end_of_stream();
-      } else {
-        EXPECT_TRUE(body_request.has_request_body());
-        body_received = absl::StrCat(body_received, body_request.request_body().body());
-        end_stream = body_request.request_body().end_of_stream();
-      }
-      total_req_body_msg++;
-    }
-    EXPECT_TRUE(end_stream);
-    if (compare_body) {
-      EXPECT_EQ(body_received, body_sent);
-    }
-    return total_req_body_msg;
-  }
-
-  void serverSendHeaderRespDuplexStreamed(bool first_message = true, bool response = false) {
-    if (first_message) {
-      processor_stream_->startGrpcStream();
-    }
-    ProcessingResponse response_header;
-    HeadersResponse* header_resp;
-    if (response) {
-      header_resp = response_header.mutable_response_headers();
-    } else {
-      header_resp = response_header.mutable_request_headers();
-    }
-    auto* header_mutation = header_resp->mutable_response()->mutable_header_mutation();
-    auto* sh = header_mutation->add_set_headers();
-    auto* header = sh->mutable_header();
-    sh->mutable_append()->set_value(false);
-    header->set_key("x-new-header");
-    header->set_raw_value("new");
-    processor_stream_->sendGrpcMessage(response_header);
-  }
-
-  void serverSendBodyRespDuplexStreamed(uint32_t total_resp_body_msg, bool end_of_stream = true,
-                                        bool response = false, absl::string_view body_sent = "") {
-    for (uint32_t i = 0; i < total_resp_body_msg; i++) {
-      ProcessingResponse response_body;
-      BodyResponse* body_resp;
-      if (response) {
-        body_resp = response_body.mutable_response_body();
-      } else {
-        body_resp = response_body.mutable_request_body();
-      }
-
-      auto* body_mut = body_resp->mutable_response()->mutable_body_mutation();
-      auto* streamed_response = body_mut->mutable_streamed_response();
-      if (!body_sent.empty()) {
-        streamed_response->set_body(body_sent);
-      } else {
-        streamed_response->set_body("r");
-      }
-      if (end_of_stream) {
-        const bool end_of_stream = (i == total_resp_body_msg - 1) ? true : false;
-        streamed_response->set_end_of_stream(end_of_stream);
-      }
-      processor_stream_->sendGrpcMessage(response_body);
-    }
-  }
-
-  void serverSendTrailerRespDuplexStreamed() {
-    ProcessingResponse response_trailer;
-    auto* trailer_resp = response_trailer.mutable_request_trailers()->mutable_header_mutation();
-    auto* sh = trailer_resp->add_set_headers();
-    sh->mutable_append()->set_value(false);
-    auto* header = sh->mutable_header();
-    header->set_key("x-new-trailer");
-    header->set_raw_value("new");
-    processor_stream_->sendGrpcMessage(response_trailer);
-  }
-
-  std::unique_ptr<SimpleFilterConfig<DynamicMetadataToHeadersFilter>> simple_filter_config_;
-  std::unique_ptr<
-      Envoy::Registry::InjectFactory<Server::Configuration::NamedHttpFilterConfigFactory>>
-      registration_;
-  std::unique_ptr<TestOnProcessingResponseFactory> processing_response_factory_;
-  std::unique_ptr<Envoy::Registry::InjectFactory<OnProcessingResponseFactory>>
-      processing_response_factory_registration_;
-  envoy::extensions::filters::http::ext_proc::v3::ExternalProcessor proto_config_{};
-  bool protocol_config_encoded_ = false;
-  ProtocolConfiguration protocol_config_{};
-  uint32_t max_message_timeout_ms_{0};
-  std::vector<FakeUpstream*> grpc_upstreams_;
-  FakeHttpConnectionPtr processor_connection_;
-  FakeStreamPtr processor_stream_;
-  TestScopedRuntime scoped_runtime_;
-  // Number of grpc upstreams in the test.
-  int grpc_upstream_count_ = 2;
-};
 
 INSTANTIATE_TEST_SUITE_P(IpVersionsClientTypeDeferredProcessing, ExtProcIntegrationTest,
                          GRPC_CLIENT_INTEGRATION_PARAMS);
@@ -4426,34 +3556,6 @@ TEST_P(ExtProcIntegrationTest, SendAndReceiveDynamicMetadata) {
   verifyDownstreamResponse(*response, 200);
 }
 
-TEST_P(ExtProcIntegrationTest, SendAndReceiveDynamicMetadataObservabilityMode) {
-  proto_config_.set_observability_mode(true);
-  proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SEND);
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-
-  auto* md_opts = proto_config_.mutable_metadata_options();
-  md_opts->mutable_forwarding_namespaces()->add_untyped("forwarding_ns_untyped");
-  md_opts->mutable_forwarding_namespaces()->add_typed("forwarding_ns_typed");
-  md_opts->mutable_receiving_namespaces()->add_untyped("receiving_ns_untyped");
-
-  ConfigOptions config_option = {};
-  config_option.add_metadata = true;
-  initializeConfig(config_option);
-  HttpIntegrationTest::initialize();
-
-  auto response = sendDownstreamRequest(absl::nullopt);
-
-  testSendDyanmicMetadata();
-
-  handleUpstreamRequest();
-
-  ASSERT_TRUE(response->waitForEndStream());
-  ASSERT_TRUE(response->complete());
-  // No headers from dynamic metadata response as the response is ignored in observability mode.
-  EXPECT_THAT(response->headers(), Not(ContainsHeader("receiving_ns_untyped.foo", _)));
-  verifyDownstreamResponse(*response, 200);
-}
-
 #if defined(USE_CEL_PARSER)
 TEST_P(ExtProcIntegrationTest, RequestResponseAttributes) {
   proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SEND);
@@ -4741,433 +3843,6 @@ TEST_P(ExtProcIntegrationTest, RetryOnDifferentHost) {
                 .getStringView(),
             "bluh");
   verifyDownstreamResponse(*response, 200);
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithHeader) {
-  proto_config_.set_observability_mode(true);
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-
-  auto response = sendDownstreamRequest(absl::nullopt);
-  Http::TestRequestHeaderMapImpl expected_request_headers{{":scheme", "http"},
-                                                          {":method", "GET"},
-                                                          {"host", "host"},
-                                                          {":path", "/"},
-                                                          {"x-forwarded-proto", "http"}};
-  processRequestHeadersMessage(
-      *grpc_upstreams_[0], true,
-      [&expected_request_headers](const HttpHeaders& headers, HeadersResponse& headers_resp) {
-        // Verify the header request.
-        EXPECT_THAT(headers.headers(), HeaderProtosEqual(expected_request_headers));
-        EXPECT_TRUE(headers.end_of_stream());
-
-        // Try to mutate the header.
-        auto* response_mutation = headers_resp.mutable_response()->mutable_header_mutation();
-        auto* add1 = response_mutation->add_set_headers();
-        add1->mutable_append()->set_value(false);
-        add1->mutable_header()->set_key("x-response-processed");
-        add1->mutable_header()->set_raw_value("1");
-        return true;
-      });
-
-  ASSERT_TRUE(fake_upstreams_[0]->waitForHttpConnection(*dispatcher_, fake_upstream_connection_));
-  ASSERT_TRUE(fake_upstream_connection_->waitForNewStream(*dispatcher_, upstream_request_));
-  ASSERT_TRUE(upstream_request_->waitForEndStream(*dispatcher_));
-  // Header mutation response has been ignored.
-  EXPECT_THAT(upstream_request_->headers(), Not(ContainsHeader("x-remove-this", _)));
-
-  Http::TestResponseHeaderMapImpl response_headers =
-      Http::TestResponseHeaderMapImpl{{":status", "200"}};
-  upstream_request_->encodeHeaders(response_headers, false);
-  upstream_request_->encodeData(100, true);
-
-  processResponseHeadersMessage(
-      *grpc_upstreams_[0], false, [](const HttpHeaders& headers, HeadersResponse&) {
-        Http::TestRequestHeaderMapImpl expected_response_headers{{":status", "200"}};
-        EXPECT_THAT(headers.headers(), HeaderProtosEqual(expected_response_headers));
-        return true;
-      });
-
-  verifyDownstreamResponse(*response, 200);
-
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS));
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithBody) {
-  proto_config_.set_observability_mode(true);
-
-  proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SKIP);
-  proto_config_.mutable_processing_mode()->set_request_body_mode(ProcessingMode::STREAMED);
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-  proto_config_.mutable_processing_mode()->set_response_body_mode(ProcessingMode::STREAMED);
-
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-  const std::string original_body_str = "Hello";
-  auto response = sendDownstreamRequestWithBody(original_body_str, absl::nullopt);
-
-  processRequestBodyMessage(*grpc_upstreams_[0], true,
-                            [&original_body_str](const HttpBody& body, BodyResponse& body_resp) {
-                              // Verify the received body message.
-                              EXPECT_EQ(body.body(), original_body_str);
-                              EXPECT_TRUE(body.end_of_stream());
-                              // Try to mutate the body.
-                              auto* body_mut =
-                                  body_resp.mutable_response()->mutable_body_mutation();
-                              body_mut->set_body("Hello, World!");
-                              return true;
-                            });
-
-  ASSERT_TRUE(fake_upstreams_[0]->waitForHttpConnection(*dispatcher_, fake_upstream_connection_));
-  ASSERT_TRUE(fake_upstream_connection_->waitForNewStream(*dispatcher_, upstream_request_));
-  ASSERT_TRUE(upstream_request_->waitForEndStream(*dispatcher_));
-  // Body mutation response has been ignored.
-  EXPECT_EQ(upstream_request_->body().toString(), original_body_str);
-
-  Http::TestResponseHeaderMapImpl response_headers =
-      Http::TestResponseHeaderMapImpl{{":status", "200"}};
-  upstream_request_->encodeHeaders(response_headers, false);
-  upstream_request_->encodeData(100, true);
-
-  processResponseBodyMessage(*grpc_upstreams_[0], false, [](const HttpBody& body, BodyResponse&) {
-    EXPECT_TRUE(body.end_of_stream());
-    return true;
-  });
-
-  verifyDownstreamResponse(*response, 200);
-
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS));
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithWrongBodyMode) {
-  proto_config_.set_observability_mode(true);
-
-  proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SKIP);
-  proto_config_.mutable_processing_mode()->set_request_body_mode(ProcessingMode::BUFFERED);
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-  const std::string original_body_str = "Hello";
-  auto response = sendDownstreamRequestWithBody(original_body_str, absl::nullopt);
-
-  handleUpstreamRequest();
-  verifyDownstreamResponse(*response, 200);
-
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS));
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithTrailer) {
-  proto_config_.set_observability_mode(true);
-
-  proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SKIP);
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-  proto_config_.mutable_processing_mode()->set_response_trailer_mode(ProcessingMode::SEND);
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequest(absl::nullopt);
-
-  handleUpstreamRequestWithTrailer();
-
-  processResponseTrailersMessage(
-      *grpc_upstreams_[0], true, [](const HttpTrailers& trailers, TrailersResponse& resp) {
-        // Verify the trailer
-        Http::TestResponseTrailerMapImpl expected_trailers{{"x-test-trailers", "Yes"}};
-        EXPECT_THAT(trailers.trailers(), HeaderProtosEqual(expected_trailers));
-
-        // Try to mutate the trailer
-        auto* trailer_mut = resp.mutable_header_mutation();
-        auto* trailer_add = trailer_mut->add_set_headers();
-        trailer_add->mutable_append()->set_value(false);
-        trailer_add->mutable_header()->set_key("x-modified-trailers");
-        trailer_add->mutable_header()->set_raw_value("xxx");
-        return true;
-      });
-
-  verifyDownstreamResponse(*response, 200);
-  EXPECT_THAT(*(response->trailers()), Not(ContainsHeader("x-modified-trailers", _)));
-
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS));
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithFullRequest) {
-  proto_config_.set_observability_mode(true);
-  uint32_t deferred_close_timeout_ms = 1000;
-  proto_config_.mutable_deferred_close_timeout()->set_seconds(deferred_close_timeout_ms / 1000);
-
-  proto_config_.mutable_processing_mode()->set_request_body_mode(ProcessingMode::STREAMED);
-  proto_config_.mutable_processing_mode()->set_request_trailer_mode(ProcessingMode::SEND);
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-  const std::string body_str = "Hello";
-  auto response = sendDownstreamRequestWithBodyAndTrailer(body_str);
-
-  processRequestHeadersMessage(*grpc_upstreams_[0], true, absl::nullopt);
-  processRequestBodyMessage(*grpc_upstreams_[0], false, absl::nullopt);
-  processRequestTrailersMessage(*grpc_upstreams_[0], false, absl::nullopt);
-
-  handleUpstreamRequest();
-  verifyDownstreamResponse(*response, 200);
-
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(deferred_close_timeout_ms));
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithFullResponse) {
-  proto_config_.set_observability_mode(true);
-  uint32_t deferred_close_timeout_ms = 1000;
-  proto_config_.mutable_deferred_close_timeout()->set_seconds(deferred_close_timeout_ms / 1000);
-
-  proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SKIP);
-  proto_config_.mutable_processing_mode()->set_response_body_mode(ProcessingMode::STREAMED);
-  proto_config_.mutable_processing_mode()->set_response_trailer_mode(ProcessingMode::SEND);
-
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequest(absl::nullopt);
-
-  handleUpstreamRequestWithTrailer();
-
-  processResponseHeadersMessage(*grpc_upstreams_[0], true, absl::nullopt);
-  processResponseBodyMessage(*grpc_upstreams_[0], false, absl::nullopt);
-  processResponseTrailersMessage(*grpc_upstreams_[0], false, absl::nullopt);
-
-  verifyDownstreamResponse(*response, 200);
-
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(deferred_close_timeout_ms));
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithFullRequestAndTimeout) {
-  proto_config_.set_observability_mode(true);
-  uint32_t deferred_close_timeout_ms = 2000;
-  proto_config_.mutable_deferred_close_timeout()->set_seconds(deferred_close_timeout_ms / 1000);
-
-  proto_config_.mutable_processing_mode()->set_request_body_mode(ProcessingMode::STREAMED);
-  proto_config_.mutable_processing_mode()->set_request_trailer_mode(ProcessingMode::SEND);
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequestWithBodyAndTrailer("Hello");
-
-  processRequestHeadersMessage(*grpc_upstreams_[0], true,
-                               [this](const HttpHeaders&, HeadersResponse&) {
-                                 // Advance 400 ms. Default timeout is 200ms
-                                 timeSystem().advanceTimeWaitImpl(400ms);
-                                 return false;
-                               });
-  processRequestBodyMessage(*grpc_upstreams_[0], false, [this](const HttpBody&, BodyResponse&) {
-    // Advance 400 ms. Default timeout is 200ms
-    timeSystem().advanceTimeWaitImpl(400ms);
-    return false;
-  });
-  processRequestTrailersMessage(*grpc_upstreams_[0], false,
-                                [this](const HttpTrailers&, TrailersResponse&) {
-                                  // Advance 400 ms. Default timeout is 200ms
-                                  timeSystem().advanceTimeWaitImpl(400ms);
-                                  return false;
-                                });
-
-  handleUpstreamRequest();
-  verifyDownstreamResponse(*response, 200);
-
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(deferred_close_timeout_ms - 1200));
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithLogging) {
-  proto_config_.set_observability_mode(true);
-
-  ConfigOptions config_option = {};
-  config_option.add_logging_filter = true;
-  initializeConfig(config_option);
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequest(
-      [](Http::HeaderMap& headers) { headers.addCopy(LowerCaseString("x-remove-this"), "yes"); });
-
-  processRequestHeadersMessage(*grpc_upstreams_[0], true, absl::nullopt);
-  handleUpstreamRequest();
-  processResponseHeadersMessage(*grpc_upstreams_[0], false, absl::nullopt);
-  verifyDownstreamResponse(*response, 200);
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithLoggingFailStream) {
-  proto_config_.set_observability_mode(true);
-
-  ConfigOptions config_option = {};
-  config_option.add_logging_filter = true;
-  initializeConfig(config_option);
-  testGetAndFailStream();
-}
-
-TEST_P(ExtProcIntegrationTest, ObservabilityModeWithLoggingCloseStream) {
-  proto_config_.set_observability_mode(true);
-
-  ConfigOptions config_option = {};
-  config_option.add_logging_filter = true;
-  initializeConfig(config_option);
-  testGetAndCloseStream();
-}
-
-TEST_P(ExtProcIntegrationTest, GetAndSetHeadersUpstreamObservabilityMode) {
-  proto_config_.set_observability_mode(true);
-
-  ConfigOptions config_option = {};
-  config_option.downstream_filter = false;
-
-  initializeConfig(config_option);
-  // Add ext_proc as upstream filter.
-  config_helper_.addConfigModifier([this](envoy::config::bootstrap::v3::Bootstrap& bootstrap) {
-    auto* static_resources = bootstrap.mutable_static_resources();
-    // Retrieve cluster_0.
-    auto* cluster = static_resources->mutable_clusters(0);
-    ConfigHelper::HttpProtocolOptions old_protocol_options;
-    if (cluster->typed_extension_protocol_options().contains(
-            "envoy.extensions.upstreams.http.v3.HttpProtocolOptions")) {
-      old_protocol_options = MessageUtil::anyConvert<ConfigHelper::HttpProtocolOptions>(
-          (*cluster->mutable_typed_extension_protocol_options())
-              ["envoy.extensions.upstreams.http.v3.HttpProtocolOptions"]);
-    }
-    if (old_protocol_options.http_filters().empty()) {
-      auto* http_filter = old_protocol_options.add_http_filters();
-      http_filter->set_name("envoy.filters.http.upstream_codec");
-      http_filter->mutable_typed_config()->PackFrom(
-          envoy::extensions::filters::http::upstream_codec::v3::UpstreamCodec::default_instance());
-    }
-    auto* ext_proc_filter = old_protocol_options.add_http_filters();
-    ext_proc_filter->set_name("envoy.filters.http.ext_proc");
-    ext_proc_filter->mutable_typed_config()->PackFrom(proto_config_);
-    for (int i = old_protocol_options.http_filters_size() - 1; i > 0; --i) {
-      old_protocol_options.mutable_http_filters()->SwapElements(i, i - 1);
-    }
-    (*cluster->mutable_typed_extension_protocol_options())
-        ["envoy.extensions.upstreams.http.v3.HttpProtocolOptions"]
-            .PackFrom(old_protocol_options);
-  });
-  HttpIntegrationTest::initialize();
-
-  auto response = sendDownstreamRequest(
-      [](Http::HeaderMap& headers) { headers.addCopy(LowerCaseString("x-remove-this"), "yes"); });
-
-  processRequestHeadersMessage(
-      *grpc_upstreams_[0], true, [](const HttpHeaders& headers, HeadersResponse& headers_resp) {
-        Http::TestRequestHeaderMapImpl expected_request_headers{
-            {":scheme", "http"}, {":method", "GET"},       {":authority", "host"},
-            {":path", "/"},      {"x-remove-this", "yes"}, {"x-forwarded-proto", "http"}};
-        EXPECT_THAT(headers.headers(), HeaderProtosEqual(expected_request_headers));
-        auto response_header_mutation = headers_resp.mutable_response()->mutable_header_mutation();
-        auto* mut1 = response_header_mutation->add_set_headers();
-        mut1->mutable_append()->set_value(false);
-        mut1->mutable_header()->set_key("x-new-header");
-        mut1->mutable_header()->set_raw_value("new");
-        response_header_mutation->add_remove_headers("x-remove-this");
-        return true;
-      });
-
-  ASSERT_TRUE(fake_upstreams_[0]->waitForHttpConnection(*dispatcher_, fake_upstream_connection_));
-  ASSERT_TRUE(fake_upstream_connection_->waitForNewStream(*dispatcher_, upstream_request_));
-  ASSERT_TRUE(upstream_request_->waitForEndStream(*dispatcher_));
-
-  upstream_request_->encodeHeaders(Http::TestResponseHeaderMapImpl{{":status", "200"}}, false);
-  upstream_request_->encodeData(100, true);
-
-  processResponseHeadersMessage(
-      *grpc_upstreams_[0], false, [](const HttpHeaders& headers, HeadersResponse&) {
-        Http::TestRequestHeaderMapImpl expected_response_headers{{":status", "200"}};
-        EXPECT_THAT(headers.headers(), HeaderProtosEqual(expected_response_headers));
-        return true;
-      });
-  verifyDownstreamResponse(*response, 200);
-}
-
-// Upstream filter chain is in alpha mode and it is not actively used in ext_proc at the moment.
-TEST_P(ExtProcIntegrationTest, DISABLED_GetAndSetHeadersUpstreamObservabilityModeWithLogging) {
-  proto_config_.set_observability_mode(true);
-
-  ConfigOptions config_option = {};
-  config_option.add_logging_filter = true;
-  config_option.downstream_filter = false;
-
-  initializeConfig(config_option);
-  // Add ext_proc as upstream filter.
-  config_helper_.addConfigModifier([this](envoy::config::bootstrap::v3::Bootstrap& bootstrap) {
-    auto* static_resources = bootstrap.mutable_static_resources();
-    // Retrieve cluster_0.
-    auto* cluster = static_resources->mutable_clusters(0);
-    ConfigHelper::HttpProtocolOptions old_protocol_options;
-    if (cluster->typed_extension_protocol_options().contains(
-            "envoy.extensions.upstreams.http.v3.HttpProtocolOptions")) {
-      old_protocol_options = MessageUtil::anyConvert<ConfigHelper::HttpProtocolOptions>(
-          (*cluster->mutable_typed_extension_protocol_options())
-              ["envoy.extensions.upstreams.http.v3.HttpProtocolOptions"]);
-    }
-    if (old_protocol_options.http_filters().empty()) {
-      auto* http_filter = old_protocol_options.add_http_filters();
-      http_filter->set_name("envoy.filters.http.upstream_codec");
-      http_filter->mutable_typed_config()->PackFrom(
-          envoy::extensions::filters::http::upstream_codec::v3::UpstreamCodec::default_instance());
-    }
-    auto* ext_proc_filter = old_protocol_options.add_http_filters();
-    ext_proc_filter->set_name("envoy.filters.http.ext_proc");
-    ext_proc_filter->mutable_typed_config()->PackFrom(proto_config_);
-    for (int i = old_protocol_options.http_filters_size() - 1; i > 0; --i) {
-      old_protocol_options.mutable_http_filters()->SwapElements(i, i - 1);
-    }
-    (*cluster->mutable_typed_extension_protocol_options())
-        ["envoy.extensions.upstreams.http.v3.HttpProtocolOptions"]
-            .PackFrom(old_protocol_options);
-  });
-  HttpIntegrationTest::initialize();
-
-  auto response = sendDownstreamRequest(
-      [](Http::HeaderMap& headers) { headers.addCopy(LowerCaseString("x-remove-this"), "yes"); });
-
-  processRequestHeadersMessage(
-      *grpc_upstreams_[0], true, [](const HttpHeaders& headers, HeadersResponse& headers_resp) {
-        Http::TestRequestHeaderMapImpl expected_request_headers{
-            {":scheme", "http"}, {":method", "GET"},       {":authority", "host"},
-            {":path", "/"},      {"x-remove-this", "yes"}, {"x-forwarded-proto", "http"}};
-        EXPECT_THAT(headers.headers(), HeaderProtosEqual(expected_request_headers));
-        auto response_header_mutation = headers_resp.mutable_response()->mutable_header_mutation();
-        auto* mut1 = response_header_mutation->add_set_headers();
-        mut1->mutable_append()->set_value(false);
-        mut1->mutable_header()->set_key("x-new-header");
-        mut1->mutable_header()->set_raw_value("new");
-        response_header_mutation->add_remove_headers("x-remove-this");
-        return true;
-      });
-
-  ASSERT_TRUE(fake_upstreams_[0]->waitForHttpConnection(*dispatcher_, fake_upstream_connection_));
-  ASSERT_TRUE(fake_upstream_connection_->waitForNewStream(*dispatcher_, upstream_request_));
-  ASSERT_TRUE(upstream_request_->waitForEndStream(*dispatcher_));
-
-  upstream_request_->encodeHeaders(Http::TestResponseHeaderMapImpl{{":status", "200"}}, false);
-  upstream_request_->encodeData(100, true);
-
-  processResponseHeadersMessage(
-      *grpc_upstreams_[0], false, [](const HttpHeaders& headers, HeadersResponse&) {
-        Http::TestRequestHeaderMapImpl expected_response_headers{{":status", "200"}};
-        EXPECT_THAT(headers.headers(), HeaderProtosEqual(expected_response_headers));
-        return true;
-      });
-  verifyDownstreamResponse(*response, 200);
-}
-
-TEST_P(ExtProcIntegrationTest, InvalidServerOnResponseInObservabilityMode) {
-  proto_config_.set_observability_mode(true);
-  proto_config_.mutable_processing_mode()->set_request_header_mode(ProcessingMode::SKIP);
-  proto_config_.mutable_processing_mode()->set_response_body_mode(ProcessingMode::STREAMED);
-
-  ConfigOptions config_option = {};
-  config_option.valid_grpc_server = false;
-  initializeConfig(config_option);
-  HttpIntegrationTest::initialize();
-
-  auto response = sendDownstreamRequestWithBody("Replace this!", absl::nullopt);
-  handleUpstreamRequest();
-  EXPECT_FALSE(grpc_upstreams_[0]->waitForHttpConnection(*dispatcher_, processor_connection_,
-                                                         std::chrono::milliseconds(25000)));
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS));
 }
 
 TEST_P(ExtProcIntegrationTest, SidestreamPushbackDownstream) {
@@ -5501,297 +4176,6 @@ TEST_P(ExtProcIntegrationTest, SendHeaderBodyNotSendTrailerTest) {
   verifyDownstreamResponse(*response, 200);
 }
 
-TEST_P(ExtProcIntegrationTest, ServerWaitForBodyBeforeSendsHeaderRespDuplexStreamed) {
-  const std::string body_sent(64 * 1024, 's');
-  IntegrationStreamDecoderPtr response = initAndSendDataDuplexStreamedMode(body_sent, true);
-
-  // The ext_proc server receives the headers.
-  ProcessingRequest header_request;
-  serverReceiveHeaderDuplexStreamed(header_request);
-  // The ext_proc server receives the body.
-  uint32_t total_req_body_msg = serverReceiveBodyDuplexStreamed(body_sent);
-
-  // The ext_proc server sends back the header response.
-  serverSendHeaderRespDuplexStreamed();
-  // The ext_proc server sends back the body response.
-  uint32_t total_resp_body_msg = 2 * total_req_body_msg;
-  const std::string body_upstream(total_resp_body_msg, 'r');
-  serverSendBodyRespDuplexStreamed(total_resp_body_msg);
-
-  handleUpstreamRequest();
-  EXPECT_THAT(upstream_request_->headers(), ContainsHeader("x-new-header", "new"));
-  EXPECT_EQ(upstream_request_->body().toString(), body_upstream);
-  verifyDownstreamResponse(*response, 200);
-}
-
-TEST_P(ExtProcIntegrationTest, LargeBodyTestDuplexStreamed) {
-  const std::string body_sent(2 * 1024 * 1024, 's');
-  initializeConfigDuplexStreamed(false);
-
-  // Sends 30 consecutive request, each carrying 2MB data.
-  for (int i = 0; i < 30; i++) {
-    codec_client_ = makeHttpConnection(lookupPort("http"));
-    Http::TestRequestHeaderMapImpl default_headers;
-    HttpTestUtility::addDefaultHeaders(default_headers);
-
-    std::pair<Http::RequestEncoder&, IntegrationStreamDecoderPtr> encoder_decoder =
-        codec_client_->startRequest(default_headers);
-    request_encoder_ = &encoder_decoder.first;
-    IntegrationStreamDecoderPtr response = std::move(encoder_decoder.second);
-    codec_client_->sendData(*request_encoder_, body_sent, true);
-    // The ext_proc server receives the headers.
-    ProcessingRequest header_request;
-    serverReceiveHeaderDuplexStreamed(header_request);
-    // The ext_proc server receives the body.
-    uint32_t total_req_body_msg = serverReceiveBodyDuplexStreamed(body_sent);
-    EXPECT_GT(total_req_body_msg, 0);
-    // The ext_proc server sends back the header response.
-    serverSendHeaderRespDuplexStreamed();
-    // The ext_proc server sends back body responses, which include 50 chunks,
-    // and each chunk contains 64KB data, thus totally ~3MB per request.
-    uint32_t total_resp_body_msg = 50;
-    const std::string body_response(64 * 1024, 'r');
-    const std::string body_upstream(total_resp_body_msg * 64 * 1024, 'r');
-    serverSendBodyRespDuplexStreamed(total_resp_body_msg, /*end_of_stream*/ true,
-                                     /*response*/ false, body_response);
-
-    handleUpstreamRequest();
-    EXPECT_THAT(upstream_request_->headers(), ContainsHeader("x-new-header", "new"));
-    EXPECT_EQ(upstream_request_->body().toString(), body_upstream);
-    verifyDownstreamResponse(*response, 200);
-    TearDown();
-  }
-}
-
-// Buffer the whole message including header, body and trailer before sending response.
-TEST_P(ExtProcIntegrationTest,
-       ServerWaitForBodyAndTrailerBeforeSendsHeaderRespDuplexStreamedSmallBody) {
-  const std::string body_sent(128 * 1024, 's');
-  IntegrationStreamDecoderPtr response = initAndSendDataDuplexStreamedMode(body_sent, false);
-  Http::TestRequestTrailerMapImpl request_trailers{{"x-trailer-foo", "yes"}};
-  codec_client_->sendTrailers(*request_encoder_, request_trailers);
-
-  // The ext_proc server receives the headers.
-  ProcessingRequest header_request;
-  serverReceiveHeaderDuplexStreamed(header_request);
-
-  std::string body_received;
-  bool end_stream = false;
-  uint32_t total_req_body_msg = 0;
-  while (!end_stream) {
-    ProcessingRequest request;
-    EXPECT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    EXPECT_TRUE(request.has_request_body() || request.has_request_trailers());
-    if (!request.has_request_trailers()) {
-      // request_body is received
-      body_received = absl::StrCat(body_received, request.request_body().body());
-      total_req_body_msg++;
-    } else {
-      // request_trailer is received.
-      end_stream = true;
-    }
-  }
-  EXPECT_TRUE(end_stream);
-  EXPECT_EQ(body_received, body_sent);
-
-  // The ext_proc server sends back the header response.
-  serverSendHeaderRespDuplexStreamed();
-
-  // The ext_proc server sends back the body response.
-  uint32_t total_resp_body_msg = total_req_body_msg / 2;
-  const std::string body_upstream(total_resp_body_msg, 'r');
-  serverSendBodyRespDuplexStreamed(total_resp_body_msg, false);
-
-  // The ext_proc server sends back the trailer response.
-  serverSendTrailerRespDuplexStreamed();
-
-  handleUpstreamRequest();
-  EXPECT_THAT(upstream_request_->headers(), ContainsHeader("x-new-header", "new"));
-  EXPECT_EQ(upstream_request_->body().toString(), body_upstream);
-  verifyDownstreamResponse(*response, 200);
-}
-
-// The body is large. The server sends some body responses after buffering some amount of data.
-// The server continuously does so until the entire body processing is done.
-TEST_P(ExtProcIntegrationTest, ServerSendBodyRespWithouRecvEntireBodyDuplexStreamed) {
-  const std::string body_sent(256 * 1024, 's');
-  IntegrationStreamDecoderPtr response = initAndSendDataDuplexStreamedMode(body_sent, false);
-  Http::TestRequestTrailerMapImpl request_trailers{{"x-trailer-foo", "yes"}};
-  codec_client_->sendTrailers(*request_encoder_, request_trailers);
-
-  // The ext_proc server receives the headers.
-  ProcessingRequest header_request;
-  serverReceiveHeaderDuplexStreamed(header_request);
-  Http::TestRequestHeaderMapImpl expected_request_headers{{":scheme", "http"},
-                                                          {":method", "GET"},
-                                                          {"host", "host"},
-                                                          {":path", "/"},
-                                                          {"x-forwarded-proto", "http"}};
-  EXPECT_THAT(header_request.request_headers().headers(),
-              HeaderProtosEqual(expected_request_headers));
-
-  std::string body_received;
-  bool end_stream = false;
-  uint32_t total_req_body_msg = 0;
-  bool header_resp_sent = false;
-  std::string body_upstream;
-
-  while (!end_stream) {
-    ProcessingRequest request;
-    EXPECT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-    EXPECT_TRUE(request.has_request_body() || request.has_request_trailers());
-    if (!request.has_request_trailers()) {
-      // Buffer the entire body.
-      body_received = absl::StrCat(body_received, request.request_body().body());
-      total_req_body_msg++;
-      // After receiving every 7 body chunks, the server sends back three body responses.
-      if (total_req_body_msg % 7 == 0) {
-        if (!header_resp_sent) {
-          // Before sending the 1st body response, sends a header response.
-          serverSendHeaderRespDuplexStreamed();
-          header_resp_sent = true;
-        }
-        ProcessingResponse response_body;
-        for (uint32_t i = 0; i < 3; i++) {
-          body_upstream += std::to_string(i);
-          auto* streamed_response = response_body.mutable_request_body()
-                                        ->mutable_response()
-                                        ->mutable_body_mutation()
-                                        ->mutable_streamed_response();
-          streamed_response->set_body(std::to_string(i));
-          processor_stream_->sendGrpcMessage(response_body);
-        }
-      }
-    } else {
-      // request_trailer is received.
-      end_stream = true;
-      Http::TestResponseTrailerMapImpl expected_trailers{{"x-trailer-foo", "yes"}};
-      EXPECT_THAT(request.request_trailers().trailers(), HeaderProtosEqual(expected_trailers));
-    }
-  }
-  EXPECT_TRUE(end_stream);
-  EXPECT_EQ(body_received, body_sent);
-
-  // Send one more body response at the end.
-  ProcessingResponse response_body;
-  auto* streamed_response = response_body.mutable_request_body()
-                                ->mutable_response()
-                                ->mutable_body_mutation()
-                                ->mutable_streamed_response();
-  streamed_response->set_body("END");
-  processor_stream_->sendGrpcMessage(response_body);
-  body_upstream += "END";
-
-  // The ext_proc server sends back the trailer response.
-  serverSendTrailerRespDuplexStreamed();
-
-  handleUpstreamRequest();
-  EXPECT_THAT(upstream_request_->headers(), ContainsHeader("x-new-header", "new"));
-  EXPECT_EQ(upstream_request_->body().toString(), body_upstream);
-  verifyDownstreamResponse(*response, 200);
-}
-
-TEST_P(ExtProcIntegrationTest, DuplexStreamedInBothDirection) {
-  const std::string body_sent(8 * 1024, 's');
-  IntegrationStreamDecoderPtr response = initAndSendDataDuplexStreamedMode(body_sent, true, true);
-
-  // The ext_proc server receives the headers/body.
-  ProcessingRequest header_request;
-  serverReceiveHeaderDuplexStreamed(header_request);
-  uint32_t total_req_body_msg = serverReceiveBodyDuplexStreamed(body_sent);
-
-  // The ext_proc server sends back the response.
-  serverSendHeaderRespDuplexStreamed();
-  uint32_t total_resp_body_msg = 2 * total_req_body_msg;
-  const std::string body_upstream(total_resp_body_msg, 'r');
-  serverSendBodyRespDuplexStreamed(total_resp_body_msg);
-
-  handleUpstreamRequest();
-  EXPECT_THAT(upstream_request_->headers(), ContainsHeader("x-new-header", "new"));
-  EXPECT_EQ(upstream_request_->body().toString(), body_upstream);
-
-  // The ext_proc server receives the responses from backend server.
-  ProcessingRequest header_response;
-  serverReceiveHeaderDuplexStreamed(header_response, false, true);
-  uint32_t total_rsp_body_msg = serverReceiveBodyDuplexStreamed("", true, false);
-
-  // The ext_proc server sends back the response.
-  serverSendHeaderRespDuplexStreamed(false, true);
-  serverSendBodyRespDuplexStreamed(total_rsp_body_msg * 3, true, true);
-
-  verifyDownstreamResponse(*response, 200);
-}
-
-// With FULL_DUPLEX_STREAMED mode configured, failure_mode_allow can only be false.
-// If the ext_proc server sends out-of-order response, it causes Envoy to send
-// local reply to the client, and reset the HTTP stream.
-TEST_P(ExtProcIntegrationTest, ServerSendOutOfOrderResponseDuplexStreamed) {
-  const std::string body_sent(8 * 1024, 's');
-  // Enable FULL_DUPLEX_STREAMED body processing in both directions.
-  IntegrationStreamDecoderPtr response = initAndSendDataDuplexStreamedMode(body_sent, true, true);
-
-  // The ext_proc server receives the request headers and body.
-  ProcessingRequest header_request;
-  serverReceiveHeaderDuplexStreamed(header_request);
-  uint32_t total_req_body_msg = serverReceiveBodyDuplexStreamed(body_sent);
-  // The ext_proc server sends back the body response, which is wrong.
-  processor_stream_->startGrpcStream();
-  serverSendBodyRespDuplexStreamed(total_req_body_msg);
-  // Envoy sends 500 response code to the client.
-  verifyDownstreamResponse(*response, 500);
-}
-
-// The ext_proc server failed to send response in time trigger Envoy HCM stream_idle_timeout.
-TEST_P(ExtProcIntegrationTest, ServerWaitTooLongBeforeSendRespDuplexStreamed) {
-  // Set HCM stream_idle_timeout to be 10s. Note one can also set the
-  // RouteAction:idle_timeout under the route configuration to override it.
-  config_helper_.addConfigModifier(
-      [&](envoy::extensions::filters::network::http_connection_manager::v3::HttpConnectionManager&
-              hcm) -> void { hcm.mutable_stream_idle_timeout()->set_seconds(10); });
-
-  const std::string body_sent(8 * 1024, 's');
-  IntegrationStreamDecoderPtr response = initAndSendDataDuplexStreamedMode(body_sent, true);
-
-  // The ext_proc server receives the headers and body.
-  ProcessingRequest header_request;
-  serverReceiveHeaderDuplexStreamed(header_request);
-  serverReceiveBodyDuplexStreamed(body_sent);
-
-  // The ext_proc server waits for 12s before sending any response.
-  // HCM stream_idle_timeout is triggered, and local reply is sent to downstream.
-  timeSystem().advanceTimeWaitImpl(std::chrono::milliseconds(12000));
-  verifyDownstreamResponse(*response, 504);
-}
-
-// Testing the case that when the client does not send trailers, if the ext_proc server sends
-// back a synthesized trailer, it is ignored by Envoy and never reaches the upstream server.
-TEST_P(ExtProcIntegrationTest, DuplexStreamedServerResponseWithSynthesizedTrailer) {
-  const std::string body_sent(64 * 1024, 's');
-  IntegrationStreamDecoderPtr response = initAndSendDataDuplexStreamedMode(body_sent, true);
-
-  // The ext_proc server receives the headers.
-  ProcessingRequest header_request;
-  serverReceiveHeaderDuplexStreamed(header_request);
-  // The ext_proc server receives the body.
-  uint32_t total_req_body_msg = serverReceiveBodyDuplexStreamed(body_sent);
-
-  // The ext_proc server sends back the header response.
-  serverSendHeaderRespDuplexStreamed();
-  // The ext_proc server sends back the body response.
-  uint32_t total_resp_body_msg = 2 * total_req_body_msg;
-  const std::string body_upstream(total_resp_body_msg, 'r');
-  // The end_of_stream of the last body response is false.
-  serverSendBodyRespDuplexStreamed(total_resp_body_msg, false, false);
-  // The ext_proc server sends back a synthesized trailer response.
-  serverSendTrailerRespDuplexStreamed();
-
-  handleUpstreamRequest();
-  EXPECT_THAT(upstream_request_->headers(), ContainsHeader("x-new-header", "new"));
-  EXPECT_EQ(upstream_request_->body().toString(), body_upstream);
-  EXPECT_EQ(upstream_request_->trailers(), nullptr);
-  verifyDownstreamResponse(*response, 200);
-}
-
 TEST_P(ExtProcIntegrationTest, ModeOverrideAllowed) {
   proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
   proto_config_.set_allow_mode_override(true);
@@ -5890,89 +4274,6 @@ TEST_P(ExtProcIntegrationTest, RequestHeaderModeIgnoredInModeOverrideComparison)
     return true;
   });
   handleUpstreamRequest();
-  verifyDownstreamResponse(*response, 200);
-}
-
-TEST_P(ExtProcIntegrationTest, ModeOverrideNoneToFullDuplex) {
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-  proto_config_.set_allow_mode_override(true);
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-
-  std::string body_str = std::string(10, 'a');
-  std::string upstream_body_str = std::string(5, 'b');
-  auto response = sendDownstreamRequestWithBody(body_str, absl::nullopt);
-  // Process request header message.
-  processGenericMessage(
-      *grpc_upstreams_[0], true, [](const ProcessingRequest&, ProcessingResponse& resp) {
-        resp.mutable_request_headers();
-        resp.mutable_mode_override()->set_request_body_mode(ProcessingMode::FULL_DUPLEX_STREAMED);
-        return true;
-      });
-
-  processRequestBodyMessage(
-      *grpc_upstreams_[0], false,
-      [&body_str, &upstream_body_str](const HttpBody& body, BodyResponse& resp) {
-        EXPECT_TRUE(body.end_of_stream());
-        EXPECT_EQ(body.body(), body_str);
-        auto* streamed_response =
-            resp.mutable_response()->mutable_body_mutation()->mutable_streamed_response();
-        streamed_response->set_body(upstream_body_str);
-        streamed_response->set_end_of_stream(true);
-        return true;
-      });
-  handleUpstreamRequest();
-  EXPECT_EQ(upstream_request_->body().toString(), upstream_body_str);
-  verifyDownstreamResponse(*response, 200);
-}
-
-TEST_P(ExtProcIntegrationTest, NoneToFullDuplexMoreDataAfterModeOverride) {
-  proto_config_.mutable_processing_mode()->set_response_header_mode(ProcessingMode::SKIP);
-  proto_config_.set_allow_mode_override(true);
-  initializeConfig();
-  HttpIntegrationTest::initialize();
-
-  codec_client_ = makeHttpConnection(lookupPort("http"));
-  Http::TestRequestHeaderMapImpl headers;
-  HttpTestUtility::addDefaultHeaders(headers);
-  auto encoder_decoder = codec_client_->startRequest(headers);
-  request_encoder_ = &encoder_decoder.first;
-  codec_client_->sendData(*request_encoder_, 10, false);
-  IntegrationStreamDecoderPtr response = std::move(encoder_decoder.second);
-  // Process request header message.
-  processGenericMessage(
-      *grpc_upstreams_[0], true, [](const ProcessingRequest&, ProcessingResponse& resp) {
-        resp.mutable_request_headers();
-        resp.mutable_mode_override()->set_request_body_mode(ProcessingMode::FULL_DUPLEX_STREAMED);
-        return true;
-      });
-
-  processRequestBodyMessage(
-      *grpc_upstreams_[0], false, [](const HttpBody& body, BodyResponse& resp) {
-        EXPECT_FALSE(body.end_of_stream());
-        EXPECT_EQ(body.body().size(), 10);
-        auto* streamed_response =
-            resp.mutable_response()->mutable_body_mutation()->mutable_streamed_response();
-        streamed_response->set_body("bbbbb");
-        streamed_response->set_end_of_stream(false);
-        return true;
-      });
-
-  codec_client_->sendData(*request_encoder_, 20, true);
-
-  processRequestBodyMessage(
-      *grpc_upstreams_[0], false, [](const HttpBody& body, BodyResponse& resp) {
-        EXPECT_TRUE(body.end_of_stream());
-        EXPECT_EQ(body.body().size(), 20);
-        auto* streamed_response =
-            resp.mutable_response()->mutable_body_mutation()->mutable_streamed_response();
-        streamed_response->set_body("0123456789");
-        streamed_response->set_end_of_stream(true);
-        return true;
-      });
-
-  handleUpstreamRequest();
-  EXPECT_EQ(upstream_request_->body().toString(), "bbbbb0123456789");
   verifyDownstreamResponse(*response, 200);
 }
 
@@ -6225,208 +4526,7 @@ TEST_P(ExtProcIntegrationTest, ExtProcLoggingInfoGRPCTimeout) {
   EXPECT_EQ(*field_request_header_status, "4");
 }
 
-// Tests for status_on_error functionality.
-class ExtProcStatusOnErrorIntegrationTest : public HttpIntegrationTest,
-                                            public Grpc::GrpcClientIntegrationParamTest {
-protected:
-  ExtProcStatusOnErrorIntegrationTest()
-      : HttpIntegrationTest(Http::CodecType::HTTP2, ipVersion()) {}
-
-  void createUpstreams() override {
-    HttpIntegrationTest::createUpstreams();
-
-    // Create separate "upstreams" for ExtProc gRPC servers
-    for (int i = 0; i < grpc_upstream_count_; ++i) {
-      grpc_upstreams_.push_back(&addFakeUpstream(Http::CodecType::HTTP2));
-    }
-  }
-
-  void TearDown() override {
-    if (processor_connection_) {
-      ASSERT_TRUE(processor_connection_->close());
-      ASSERT_TRUE(processor_connection_->waitForDisconnect());
-    }
-    cleanupUpstreamAndDownstream();
-  }
-
-  void initializeConfig(uint32_t status_code) {
-    config_helper_.addConfigModifier([this, status_code](
-                                         envoy::config::bootstrap::v3::Bootstrap& bootstrap) {
-      auto* server_cluster = bootstrap.mutable_static_resources()->add_clusters();
-      server_cluster->MergeFrom(bootstrap.static_resources().clusters()[0]);
-      server_cluster->set_name("ext_proc_server");
-      server_cluster->mutable_load_assignment()->set_cluster_name("ext_proc_server");
-
-      setGrpcService(*proto_config_.mutable_grpc_service(), "ext_proc_server",
-                     grpc_upstreams_[0]->localAddress());
-
-      proto_config_.mutable_status_on_error()->set_code(
-          static_cast<envoy::type::v3::StatusCode>(status_code));
-      proto_config_.set_failure_mode_allow(false);
-
-      envoy::extensions::filters::network::http_connection_manager::v3::HttpFilter ext_proc_filter;
-      ext_proc_filter.set_name("envoy.filters.http.ext_proc");
-      ext_proc_filter.mutable_typed_config()->PackFrom(proto_config_);
-      config_helper_.prependFilter(MessageUtil::getJsonStringFromMessageOrError(ext_proc_filter));
-    });
-
-    setUpstreamProtocol(Http::CodecType::HTTP1);
-    setDownstreamProtocol(Http::CodecType::HTTP1);
-  }
-
-  IntegrationStreamDecoderPtr sendDownstreamRequest(
-      absl::optional<std::function<void(Http::RequestHeaderMap& headers)>> modify_headers) {
-    auto conn = makeClientConnection(lookupPort("http"));
-    codec_client_ = makeHttpConnection(std::move(conn));
-    Http::TestRequestHeaderMapImpl headers;
-    HttpTestUtility::addDefaultHeaders(headers);
-    if (modify_headers) {
-      (*modify_headers)(headers);
-    }
-    return codec_client_->makeHeaderOnlyRequest(headers);
-  }
-
-  void waitForFirstMessage(FakeUpstream& grpc_upstream, ProcessingRequest& request) {
-    ASSERT_TRUE(grpc_upstream.waitForHttpConnection(*dispatcher_, processor_connection_));
-    ASSERT_TRUE(processor_connection_->waitForNewStream(*dispatcher_, processor_stream_));
-    ASSERT_TRUE(processor_stream_->waitForGrpcMessage(*dispatcher_, request));
-  }
-
-  bool IsEnvoyGrpc() { return std::get<1>(GetParam()) == Envoy::Grpc::ClientType::EnvoyGrpc; }
-
-  envoy::extensions::filters::http::ext_proc::v3::ExternalProcessor proto_config_{};
-  FakeStreamPtr processor_stream_;
-  FakeHttpConnectionPtr processor_connection_;
-  std::vector<FakeUpstream*> grpc_upstreams_;
-  int grpc_upstream_count_ = 1;
-};
-
-INSTANTIATE_TEST_SUITE_P(IpVersionsClientType, ExtProcStatusOnErrorIntegrationTest,
-                         GRPC_CLIENT_INTEGRATION_PARAMS,
-                         Grpc::GrpcClientIntegrationParamTest::protocolTestParamsToString);
-
-// Test that status_on_error is used when gRPC stream encounters an error.
-TEST_P(ExtProcStatusOnErrorIntegrationTest, GrpcStreamErrorCustomStatus) {
-  SKIP_IF_GRPC_CLIENT(Grpc::ClientType::EnvoyGrpc);
-
-  initializeConfig(503); // Use 503 Service Unavailable.
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequest(absl::nullopt);
-
-  ProcessingRequest request_headers_msg;
-  waitForFirstMessage(*grpc_upstreams_[0], request_headers_msg);
-  processor_stream_->startGrpcStream();
-
-  // Send successful response first, then simulate gRPC stream error.
-  ProcessingResponse resp;
-  resp.mutable_request_headers();
-  processor_stream_->sendGrpcMessage(resp);
-
-  // Now trigger gRPC stream error which should use status_on_error.
-  processor_stream_->finishGrpcStream(Grpc::Status::Internal);
-
-  // Should get custom status code 503 instead of default 500.
-  ASSERT_TRUE(response->waitForEndStream());
-  EXPECT_EQ("503", response->headers().getStatusValue());
-}
-
-// Test that default status (500) is used when status_on_error is not configured.
-TEST_P(ExtProcStatusOnErrorIntegrationTest, GrpcStreamErrorDefaultStatus) {
-  SKIP_IF_GRPC_CLIENT(Grpc::ClientType::EnvoyGrpc);
-
-  // Initialize without setting status_on_error, should default to 500.
-  config_helper_.addConfigModifier([this](envoy::config::bootstrap::v3::Bootstrap& bootstrap) {
-    auto* server_cluster = bootstrap.mutable_static_resources()->add_clusters();
-    server_cluster->MergeFrom(bootstrap.static_resources().clusters()[0]);
-    server_cluster->set_name("ext_proc_server");
-    server_cluster->mutable_load_assignment()->set_cluster_name("ext_proc_server");
-
-    setGrpcService(*proto_config_.mutable_grpc_service(), "ext_proc_server",
-                   grpc_upstreams_[0]->localAddress());
-
-    proto_config_.set_failure_mode_allow(false);
-
-    envoy::extensions::filters::network::http_connection_manager::v3::HttpFilter ext_proc_filter;
-    ext_proc_filter.set_name("envoy.filters.http.ext_proc");
-    ext_proc_filter.mutable_typed_config()->PackFrom(proto_config_);
-    config_helper_.prependFilter(MessageUtil::getJsonStringFromMessageOrError(ext_proc_filter));
-  });
-
-  setUpstreamProtocol(Http::CodecType::HTTP1);
-  setDownstreamProtocol(Http::CodecType::HTTP1);
-
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequest(absl::nullopt);
-
-  ProcessingRequest request_headers_msg;
-  waitForFirstMessage(*grpc_upstreams_[0], request_headers_msg);
-  processor_stream_->startGrpcStream();
-
-  // Send successful response first, then simulate gRPC stream error.
-  ProcessingResponse resp;
-  resp.mutable_request_headers();
-  processor_stream_->sendGrpcMessage(resp);
-
-  // Trigger gRPC stream error without status_on_error configured.
-  processor_stream_->finishGrpcStream(Grpc::Status::Internal);
-
-  // Should get default status code 500.
-  ASSERT_TRUE(response->waitForEndStream());
-  EXPECT_EQ("500", response->headers().getStatusValue());
-}
-
-// Test that message timeout returns 504 Gateway Timeout regardless of status_on_error.
-TEST_P(ExtProcStatusOnErrorIntegrationTest, MessageTimeoutReturnsGatewayTimeout) {
-  SKIP_IF_GRPC_CLIENT(Grpc::ClientType::EnvoyGrpc);
-
-  initializeConfig(502); // Configure 502, but timeout should return 504.
-  proto_config_.mutable_message_timeout()->set_nanos(100000000); // 100ms timeout.
-
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequest(absl::nullopt);
-
-  ProcessingRequest request_headers_msg;
-  waitForFirstMessage(*grpc_upstreams_[0], request_headers_msg);
-  processor_stream_->startGrpcStream();
-
-  // Don't send a response to trigger timeout - just wait for the timeout to occur.
-
-  // Let timeout occur.
-  test_server_->waitForCounterGe("http.config_test.ext_proc.message_timeouts", 1);
-
-  // Should return 504 Gateway Timeout instead of configured status_on_error.
-  ASSERT_TRUE(response->waitForEndStream());
-  EXPECT_EQ("504", response->headers().getStatusValue());
-}
-
-// Test that status_on_error is used when processing/mutation errors occur.
-TEST_P(ExtProcStatusOnErrorIntegrationTest, ProcessingErrorCustomStatus) {
-  SKIP_IF_GRPC_CLIENT(Grpc::ClientType::EnvoyGrpc);
-
-  initializeConfig(422); // Use 422 Unprocessable Entity.
-  // Enable strict mutation rules to trigger processing errors.
-  proto_config_.mutable_mutation_rules()->mutable_disallow_is_error()->set_value(true);
-  proto_config_.mutable_mutation_rules()->mutable_disallow_system()->set_value(true);
-
-  HttpIntegrationTest::initialize();
-  auto response = sendDownstreamRequest(absl::nullopt);
-
-  // Process the request and send back invalid system header mutation.
-  ProcessingRequest request_headers_msg;
-  waitForFirstMessage(*grpc_upstreams_[0], request_headers_msg);
-  processor_stream_->startGrpcStream();
-
-  ProcessingResponse resp;
-  auto* header_mut = resp.mutable_request_headers()->mutable_response()->mutable_header_mutation();
-  auto* header = header_mut->add_set_headers();
-  header->mutable_append()->set_value(false);
-  header->mutable_header()->set_key(":system-header"); // This should trigger error.
-  header->mutable_header()->set_raw_value("invalid");
-  processor_stream_->sendGrpcMessage(resp);
-
-  // Should get custom status code 422 for processing error.
-  ASSERT_TRUE(response->waitForEndStream());
-  EXPECT_EQ("422", response->headers().getStatusValue());
-}
-
+} // namespace ExternalProcessing
+} // namespace HttpFilters
+} // namespace Extensions
 } // namespace Envoy
